pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markio"
version = "0.1.0"
description = "Markio Document Parser CLI and API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SimonSun"}
]
keywords = ["document", "parser", "api", "cli"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "docling>=2.37.0",
    "fastapi>=0.115.13",
    "fastapi-mcp>=0.3.4",
    "gradio>=4.0.0",
    "mineru[all]>=2.1.0",
    "pypandoc>=1.15",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "typer>=0.16.0",
    "uvicorn>=0.34.3",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
]

[project.urls]
Homepage = "https://github.com/Tendo33/markio"
Repository = "https://github.com/Tendo33/markio"
Issues = "https://github.com/Tendo33/markio/issues"

[project.scripts]
markio = "markio.sdk.markio_cli:app"

[tool.hatch.build.targets.wheel]
packages = ["markio"]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.hatch.build.targets.sdist]
include = [
    "/markio",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
